VP1-203-CleanCannonAndSurvey
Cleaned cannon and survey.
