+++ VLAT_Demo/Assets/DemoScene.unity
@@ -5123,7 +5123,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 604804290}
-  - component: {fileID: 604804291}
   m_Layer: 0
   m_Name: Interactables
   m_TagString: Untagged
@@ -5150,18 +5149,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &604804291
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 604804289}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f32149c5c37f1ef4aa9001af626bce51, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1001 &605774406
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -5705,17 +5692,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 1050340177746238356, guid: f181592ef3fc73f4f94b68e3448916dc, type: 3}
   m_PrefabInstance: {fileID: 1906836708}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &692565286 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 3106172421161745842, guid: f181592ef3fc73f4f94b68e3448916dc, type: 3}
-  m_PrefabInstance: {fileID: 1906836708}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2567436d712d86841a4de208fe64bd95, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &694284565
 GameObject:
   m_ObjectHideFlags: 0
@@ -12456,11 +12432,11 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
+  - {fileID: 692565285}
   - {fileID: 962441329}
   - {fileID: 1393589576}
   - {fileID: 1116400514}
   - {fileID: 903751342}
-  - {fileID: 692565285}
   m_Father: {fileID: 604804290}
   m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -13786,7 +13762,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4891787282750124, guid: 7087784e228be8745af45599a5130532, type: 3}
       propertyPath: m_RootOrder
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 4891787282750124, guid: 7087784e228be8745af45599a5130532, type: 3}
       propertyPath: m_LocalPosition.x
@@ -13936,8 +13912,9 @@ GameObject:
   - component: {fileID: 1469700866}
   - component: {fileID: 1469700865}
   - component: {fileID: 1469700864}
+  - component: {fileID: 1469700868}
   m_Layer: 5
-  m_Name: BlackFadeCanvas
+  m_Name: SurveyTransitioner
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -13950,9 +13927,9 @@ RectTransform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1469700862}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0.1}
-  m_LocalScale: {x: 0.05, y: 0.05, z: 0.05}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0.030000467}
+  m_LocalScale: {x: 0.000079671154, y: 0.000079671154, z: 0.000079671154}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 1843677764}
@@ -13962,7 +13939,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: 880.4, y: 386}
+  m_SizeDelta: {x: 408.39993, y: 434.8}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!225 &1469700864
 CanvasGroup:
@@ -14026,8 +14003,8 @@ Canvas:
   m_Enabled: 1
   serializedVersion: 3
   m_RenderMode: 2
-  m_Camera: {fileID: 0}
-  m_PlaneDistance: 100
+  m_Camera: {fileID: 353113963}
+  m_PlaneDistance: 0.03
   m_PixelPerfect: 0
   m_ReceivesEvents: 1
   m_OverrideSorting: 0
@@ -14037,6 +14014,20 @@ Canvas:
   m_SortingLayerID: 0
   m_SortingOrder: 0
   m_TargetDisplay: 0
+--- !u!114 &1469700868
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1469700862}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 319632dade31dd449b5d65fb8d38e086, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  xrOrigin: {fileID: 1717954560423644721}
+  surveyInfo: {fileID: 11400000, guid: fe55ef34f3e00d847a6c802c858a7465, type: 2}
 --- !u!1001 &1472804472
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -17612,7 +17603,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 1050340177746238356, guid: f181592ef3fc73f4f94b68e3448916dc, type: 3}
       propertyPath: m_RootOrder
-      value: 4
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 1050340177746238356, guid: f181592ef3fc73f4f94b68e3448916dc, type: 3}
       propertyPath: m_LocalPosition.x
@@ -18531,7 +18522,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4891787282750124, guid: 7087784e228be8745af45599a5130532, type: 3}
       propertyPath: m_RootOrder
-      value: 2
+      value: 3
       objectReference: {fileID: 0}
     - target: {fileID: 4891787282750124, guid: 7087784e228be8745af45599a5130532, type: 3}
       propertyPath: m_LocalPosition.x
@@ -18646,7 +18637,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4891787282750124, guid: 7087784e228be8745af45599a5130532, type: 3}
       propertyPath: m_RootOrder
-      value: 1
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 4891787282750124, guid: 7087784e228be8745af45599a5130532, type: 3}
       propertyPath: m_LocalPosition.x
@@ -19471,7 +19462,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4891787282750124, guid: 7087784e228be8745af45599a5130532, type: 3}
       propertyPath: m_RootOrder
-      value: 3
+      value: 4
       objectReference: {fileID: 0}
     - target: {fileID: 4891787282750124, guid: 7087784e228be8745af45599a5130532, type: 3}
       propertyPath: m_LocalPosition.x
@@ -22227,12 +22218,6 @@ MonoBehaviour:
   cannonBody: {fileID: 1913068158673537745}
   cannonballSpawnpoint: {fileID: 5865654339940737954}
   cannonballPrefab: {fileID: 1595082824661717577, guid: d3f9bf2c465a5294d914f663b36d62eb, type: 3}
-  cannonFireSpeed: 10
-  cannonText: {fileID: 1189530269}
-  surveyManager: {fileID: 692565286}
-  fadeCanvGroup: {fileID: 1469700864}
-  xrOrigin: {fileID: 1717954560423644722}
-  exSurveyInfo: {fileID: 11400000, guid: fe55ef34f3e00d847a6c802c858a7465, type: 2}
 --- !u!33 &2321328507512280056
 MeshFilter:
   m_ObjectHideFlags: 0
