+++ VLAT_Demo/Assets/VERA/ExampleInteractables/Scripts/TransitionToSurvey.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 319632dade31dd449b5d65fb8d38e086
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
